更新(高橋)

+++ src/com/internousdev/craftdenki/dao/FavoriteDAO.java
@@ -9,7 +9,6 @@ import java.util.ArrayList;
 import com.internousdev.craftdenki.dto.FavoriteDTO;
 import com.internousdev.craftdenki.util.DBConnector;
 
-
 public class FavoriteDAO {
 
 	private DBConnector dbConnector = new DBConnector();
@@ -18,20 +17,17 @@ public class FavoriteDAO {
 	public ArrayList<FavoriteDTO> getFavoriteInfo(String loginId) throws SQLException {
 		ArrayList<FavoriteDTO> favoriteDTO = new ArrayList<FavoriteDTO>();
 		// SQL文わからない
-		String sql = "SELECT product_name, product_name_kana, image_file_path,image_file_name, price, release_company, release_date from favorite_info JOIN product_info ON favorite_info.product_id = product_info.product_id where user_id = ?";
-
+		String sql = "SELECT pi.product_id as product_id,pi.product_name as product_name,pi.product_name_kana as product_name_kana,pi.image_file_path as image_file_path,pi.image_file_name as image_file_name,pi.price as price,pi.release_company as release_company,pi.release_date as release_date from product_info as pi JOIN favorite_info as fi ON fi.product_id = pi.product_id where fi.user_id = ?";
 
 		try {
 			PreparedStatement ps = connection.prepareStatement(sql);
-
 			ps.setString(1, loginId);
-
 			ResultSet rs = ps.executeQuery();
-
-
+			System.out.println("ここまで");
 
 			while (rs.next()) {
 				FavoriteDTO dto = new FavoriteDTO();
+				dto.setProductId(rs.getString("product_id"));
 				dto.setProductName(rs.getString("product_name"));
 				dto.setProductNameKana(rs.getString("product_name_kana"));
 				dto.setImageFilePath(rs.getString("image_file_path"));
@@ -40,38 +36,31 @@ public class FavoriteDAO {
 				dto.setReleaseCompany(rs.getString("release_company"));
 				dto.setReleaseDate(rs.getString("release_date"));
 				favoriteDTO.add(dto);
-				// FavoriteDTO.add(dto);
+				System.out.println("ここまできてる");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
-		} finally {
-			connection.close();
 		}
-
 		return favoriteDTO;
 	}
 
-
-
-
-
-
-	public int deleteFavoriteInfo(int product_id){
-		DBConnector db=new DBConnector();
-		Connection con=db.getConnection();
-		int count=0;
-		String sql="delete from favorite_info where product_id = ?";
+	public int deleteFavoriteInfo(String product_id, String userid) {
+		DBConnector db = new DBConnector();
+		Connection con = db.getConnection();
+		int count = 0;
+		String sql = "delete from favorite_info where user_id = ? and product_id = ?";
 
 		try {
-			PreparedStatement ps=con.prepareStatement(sql);
-			ps.setInt(1,product_id);
+			PreparedStatement ps = con.prepareStatement(sql);
+			ps.setString(1, userid);
+			ps.setString(2, product_id);
 
 			count = ps.executeUpdate();
 			System.out.println(count);
 
-		}catch(Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
-		}finally {
+		} finally {
 			try {
 				connection.close();
 			} catch (SQLException e) {
@@ -81,40 +70,24 @@ public class FavoriteDAO {
 		return count;
 	}
 
+	// カートテーブルにInsertメソッド
+	public void insertFavorite(String user_id, String product_id) throws SQLException {
+		DBConnector dbConnector = new DBConnector();
+		Connection connection = dbConnector.getConnection();
+		String insert = "INSERT INTO favorite_info (user_id,product_id,regist_date) VALUES(?,?,NOW())";
 
+		try {
+			PreparedStatement ps = connection.prepareStatement(insert);
+			ps.setString(1, user_id);
+			ps.setString(2, product_id);
 
+			ps.execute();
 
-
-	//カートテーブルにInsertメソッド
-		public void insertFavorite(String user_id,int product_id) throws SQLException{
-			DBConnector dbConnector = new DBConnector();
-			Connection connection = dbConnector.getConnection();
-			String insert = "INSERT INTO favorite_info (user_id,product_id,regist_date) VALUES(?,?,NOW())";
-
-			try{
-				PreparedStatement ps = connection.prepareStatement(insert);
-				ps.setString(1,user_id);
-				ps.setInt(2,product_id);
-
-
-				ps.execute();
-
-			}catch(SQLException e){
-				e.printStackTrace();
-			}finally{
-				connection.close();
-			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} finally {
+			connection.close();
 		}
-
-
-
-
-
-
-
-
-
-
-
+	}
 
 }
